redpanda: split out handler get_local_offsets_translated_handler
